refactor(bot): API clientをresource別に分割する
Bot API clientの肥大化を解消するため、health/users/events/matches/match-watchers/Riot/authをresource別clientへ分離した。

共通transportと日時変換を再利用し、既存のflat facadeは移行期間の互換層として維持する。

既存のHTTP API契約と戻り値は変更せず、resource client生成のテストを追加した。
